+++ src/Java/myservicebus-abstractions/src/main/java/com/myservicebus/FilterLifetime.java
@@ -1,0 +1,7 @@
+package com.myservicebus;
+
+public enum FilterLifetime {
+    INSTANCE,
+    PIPE,
+    SCOPED
+}